pyproject.toml
[tool.poetry]
name = "eventscraper"
version = "0.1.0"
description = "Scrapes ra.co for the current weeks events and publishes them to discord"
authors = ["Jiafi <jordanwolinsky@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.28.1"
beautifulsoup4 = "^4.11.1"
"discord.py" = "^1.7.3"
selenium = "^4.5.0"
webdriver-manager = "^3.8.3"
chromedriver-autoinstaller = "^0.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
